+++ plugins/services/src/js/utils/ServiceSpecUtil.js
@@ -1,7 +1,7 @@
 import PodSpec from '../structs/PodSpec';
 import ApplicationSpec from '../structs/ApplicationSpec';
 import FrameworkSpec from '../structs/FrameworkSpec';
-import Util from './Util';
+import Util from '../../../../../src/js/utils/Util';
 
 var ServiceSpecUtil = {
